diem-bft/requirements.txt
pynacl==1.4.0
anyio==3.4.0
asgiref==3.4.1
blinker==1.4
brotlipy==0.7.0
cffi==1.15.0
click==8.0.3
decorator==5.1.0
dnslib==0.9.16
Flask==2.0.2
h11==0.12.0
httpbin==0.7.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyNaCl==1.4.0
raven==6.10.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
Werkzeug==2.0.2
wsproto==1.0.0
zipp==3.6.0
-e git+https://github.com/aiortc/aioquic.git@645dd009af08f8723a44db1d22028b75358b6176#egg=aioquic
anyio==3.4.0
asgiref==3.4.1
blinker==1.4
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cryptography==3.4.8
decorator==5.1.0
dnslib==0.9.16
Flask==2.0.2
h11==0.12.0
httpbin==0.7.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
pylsqpack==0.3.13
PyNaCl==1.4.0
raven==6.10.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
Werkzeug==2.0.2
wsproto==1.0.0
zipp==3.6.0
